src/mtdr/python/requirements_linux.txt
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
Authlib==1.6.5
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
cyclopts==3.24.0
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.2
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.2
fastmcp==2.12.4
fonttools==4.59.1
fqdn==1.5.1
grpcio==1.74.0
grpcio-tools==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.2
idna==3.10
ifaddr==0.2.0
ipykernel==6.30.1
ipympl==0.9.7
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.2.2
lazy-object-proxy==1.12.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mcp==1.16.0
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.8.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy==2.3.2
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
packaging==25.0
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.5
pathable==0.4.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
Pygments==2.19.2
pyparsing==3.2.3
pyperclip==1.11.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
pyqtgraph==0.13.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rich-rst==1.3.1
rpds-py==0.27.0
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.2
typing_extensions==4.14.1
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.37.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.1
widgetsnbextension==4.0.14
zeroconf==0.147.0
